requirements.txt
flask
numpy
pandas
scikit-learn
python-dotenv
google-generativeai